Added HR for Skandika Morpheus issue #2566 (#2567)
* Added HR for Skandika Morpheus issue #2566

Because the morpheus behaves like a mixture of the x-2000 and the wiry, I have introduced a new variable. Speed, watts and rpm are read out like the wiry, but the heart rate is read out like the x-2000. With this code the morpheus works for me. I have attached a screenshot in the issue #2566. to distinguish an x-2000 from a Morpheus it seems to be enough to pay attention to the number of characters of the bluetooth name

* Update skandikawiribike.cpp

---------

Co-authored-by: Roberto Viola <Cagnulein@gmail.com>